Update MATE Panel center-stick patches: Revision 1
In previous revisions, I was not careful enough to test that "expanding" (dynamically sized) applets such as the Window List and the MATE Dock applet were centered correctly.  Essentially, for the purposes of centering, they used to be treated as if they were always displayed at their minimum acceptable size -- in other words, they were not really centered!  This revision fixes that problem, so now docks, window lists, anything can be centered on the panel.